backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
pymysql
cryptography
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
cors
